switched to using cookies to store necessary data

+++ expense_tracker.py
@@ -2,6 +2,7 @@ from flask import Flask, request, session, render_template, redirect, url_for, j
 import mysql.connector
 import pymongo
 import os
+import jwt
 import json
 from flask_bcrypt import Bcrypt
 from flask_cors import CORS
@@ -17,8 +18,6 @@ MONGO_DB_PW = os.environ.get('MONGO_PW')
 MONGO_DB_DB = os.environ.get('MONGO_DB')
 MONGO_DB_COL = os.environ.get('MONGO_COL')
 
-
-
 db_config = {
     'host': os.environ.get('AWS_RDS_URI'),
     'user': os.environ.get('RDS_USERNAME'),
@@ -26,19 +25,35 @@ db_config = {
     'database': os.environ.get('RDS_DB_NAME'),
 }
 
-
-
 uri = f"mongodb+srv://{MONGO_DB_NAME}:{MONGO_DB_PW}@cluster0.wvhyisx.mongodb.net/?retryWrites=true&w=majority"
 client = pymongo.MongoClient(uri)
 db = client[MONGO_DB_DB]
 col = db[MONGO_DB_COL]
 
+def encode(user_id):    
+    payload = {
+        'user_id': user_id
+    }
+    return jwt.encode(payload, app.secret_key ,algorithm='HS256')
+
+def decode(payload):
+    decoded_payload = jwt.decode(payload, app.secret_key , algorithms=['HS256'])    
+    return decoded_payload['user_id']
 
 ###DEFINE INITIAL TEMPLATE ROUTES
 @app.route('/')
 def home():
-    return redirect("https://expense-tracker-landing.netlify.app/")
+    return render_template('landing.html')
 
+@app.route('/success', methods = ['GET'])
+def success():
+    if "user" not in session:
+         return redirect(url_for('home'))
+    return render_template('success.html')
+
+@app.route('/login', methods=['GET','POST'])
+def login():
+    return render_template('login.html')
 
 ###USER LOGIN FUNCTIONALITY
 @app.route('/user_login', methods=['POST','GET'])
@@ -57,17 +72,18 @@ def user_login():
         #If the given credentials are correct, user is redirected to their dashoard
         if bcrypt.check_password_hash(pw[0], password):
             cursor.execute(f"SELECT id FROM user_info WHERE userID = '{username}' LIMIT 1")
-            dbID = cursor.fetchone()
-            dbID = dbID[0]
-            session["user"] = dbID
-            return jsonify({'message': 'Login successful'}), 200
+            master_user_id = cursor.fetchone()
+            master_user_id = master_user_id[0]
+            encoded_id = encode(master_user_id)
+            return jsonify({'message': 'Login successful','encoded_id':encoded_id}), 200
         #If the credentials are incorrect the page refrehes with an error message
         else:
-            response = jsonify({'message': 'Invalid username or password'})
-            response.headers.add('Access-Control-Allow-Origin', '*')
-            return response
+            return jsonify({'message': 'Invalid username or password'}), 200
 
 
+@app.route('/signup', methods=['GET','POST'])
+def signup():
+    return render_template('signup.html')
 
 ###Primary Sign-Up option is the traditional username-password method
 @app.route('/signup_user', methods=['POST'])
@@ -110,13 +126,14 @@ def dashboard():
          return redirect(url_for('home'))
     return render_template('dashboard.html')
 
-@app.route('/get_income_v_expense', methods = ['GET'])
+@app.route('/get_income_v_expense', methods = ['POST'])
 def get_income_v_expense():
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     income = []
     expenses = []
 
-    query = f"SELECT SUM(amount) AS total_income, DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month FROM user_income WHERE user_id = {dbID} GROUP BY month ORDER BY month desc LIMIT 12;"
+    query = f"SELECT SUM(amount) AS total_income, DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month FROM user_income WHERE user_id = {master_user_id} GROUP BY month ORDER BY month desc LIMIT 12;"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -124,7 +141,7 @@ def get_income_v_expense():
     for row in data:
         income.append(dict(zip([column[0] for column in cursor.description], row)))
 
-    query = f"SELECT SUM(amount) AS total_expenses, DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month FROM user_expenses WHERE user_id = {dbID} GROUP BY month ORDER BY month desc LIMIT 12;"
+    query = f"SELECT SUM(amount) AS total_expenses, DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month FROM user_expenses WHERE user_id = {master_user_id} GROUP BY month ORDER BY month desc LIMIT 12;"
 
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
@@ -137,12 +154,13 @@ def get_income_v_expense():
     income_expense = {"income": income, "expenses": expenses}
     return jsonify(income_expense)
 
-@app.route('/get_income_breakdown', methods = ['GET'])
+@app.route('/get_income_breakdown', methods = ['POST'])
 def get_income_breakdown():
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     incomes = []
 
-    query = f"SELECT DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month, income_type, SUM(amount) AS income_type_sum FROM user_income WHERE user_id = {dbID} GROUP BY month, income_type ORDER BY month, income_type LIMIT 12;"
+    query = f"SELECT DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month, income_type, SUM(amount) AS income_type_sum FROM user_income WHERE user_id = {master_user_id} GROUP BY month, income_type ORDER BY month, income_type LIMIT 12;"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -152,12 +170,13 @@ def get_income_breakdown():
     conn.close
     return jsonify(incomes)
 
-@app.route('/get_expense_breakdown', methods = ['GET'])
+@app.route('/get_expense_breakdown', methods = ['POST'])
 def get_expense_breakdown():
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     expenses = []
 
-    query = f"SELECT DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month, expense_type, SUM(amount) AS expense_type_sum FROM user_expenses WHERE user_id = {dbID} GROUP BY month, expense_type ORDER BY month, expense_type LIMIT 12;"
+    query = f"SELECT DATE_FORMAT(STR_TO_DATE(day_month_year, '%Y-%m-%d'), '%Y-%m') AS month, expense_type, SUM(amount) AS expense_type_sum FROM user_expenses WHERE user_id = {master_user_id} GROUP BY month, expense_type ORDER BY month, expense_type LIMIT 12;"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -168,23 +187,27 @@ def get_expense_breakdown():
     conn.close
     return jsonify(expenses)
 
-@app.route('/get_budget_recent_expenses', methods = ['GET'])
+@app.route('/get_budget_recent_expenses', methods = ['POST'])
 def get_budget_recent_expenses():
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     monthly_expenses = []
 
-    budget_targets = col.find_one({"_id": dbID})
+    budget_targets = col.find_one({"_id": master_user_id})
     budget_targets = budget_targets['budget']
+    print(budget_targets)
 
-    query = f"SELECT expense_type, SUM(amount) AS total_amount FROM user_expenses WHERE user_id = {dbID} AND YEAR(day_month_year) = YEAR(CURDATE()) AND MONTH(day_month_year) = MONTH(CURDATE()) GROUP BY expense_type;"
+    query = f"SELECT expense_type, SUM(amount) AS total_amount FROM user_expenses WHERE user_id = {master_user_id} AND YEAR(day_month_year) = YEAR(CURDATE()) AND MONTH(day_month_year) = MONTH(CURDATE()) GROUP BY expense_type;"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
-    data = cursor.fetchall()    
+    data = cursor.fetchall() 
+    print(data)   
     for row in data:
         monthly_expenses.append(dict(zip([column[0] for column in cursor.description], row)))
 
     conn.close
+    print(monthly_expenses)
     return jsonify({'budget' : budget_targets, 'monthly_expenses' : monthly_expenses})
 
 
@@ -204,7 +227,8 @@ def add_income():
     incomeType = request.json.get('incomeType')
     amount = request.json.get('amount')
     date = request.json.get('date')
-    user = session['user']
+    encoded_id = request.json.get('encoded_id')
+    user = decode(encoded_id)
 
     #DATA INSERTION INTO SQL DATABSE
     conn = mysql.connector.connect(**db_config)
@@ -216,21 +240,24 @@ def add_income():
 
 ###FUNCTIONALITY TO GET THE VAROIUS INCOME TYPES THAT A USER HAS STORED IN 
 # THEIR NOSQL DOCUMENT AND RETURNS THE RESULTS TO THE JAVASCRIPT FONT-END
-@app.route('/get_income_types', methods = ['GET'])
+@app.route('/get_income_types', methods = ['POST'])
 def get_income_types():
-    dbID = session["user"]
-    incomeTypes = col.find_one({"_id": dbID})
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
+    incomeTypes = col.find_one({"_id": master_user_id})
     incomeTypes = incomeTypes['income_types']
+    print(incomeTypes)
     return jsonify({'types':incomeTypes})
 
 ###FUNCTIONALITY TO ADD A NEW INCOMETYPE TO THE USER'S NOSQL DOCUMENT
 @app.route('/add_income_type', methods = ['POST'])
 def add_income_type():
     newIncomeType = request.json.get('newIncomeType')
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
 
     #Gets all the INCOME TYPES in the current user's NoSQL document
-    incomeTypes = col.find_one({"_id": dbID})
+    incomeTypes = col.find_one({"_id": master_user_id})
     incomeTypes = incomeTypes['income_types'] 
 
     #If the income type trying to be added already exists, return an "exists" message to the JavaScript front-end
@@ -238,31 +265,33 @@ def add_income_type():
         return jsonify({'message' : 'exists'})
 
     #If the income type doesn't exist in the document, it is pushed onto user's NoSQL document
-    col.update_one({"_id" : dbID}, { "$push" : {"income_types" : newIncomeType}})
+    col.update_one({"_id" : master_user_id}, { "$push" : {"income_types" : newIncomeType}})
     return jsonify({'message' : 'success'})
 
 ### FUNCTIONALITY TO REMOVE AN INCOME TYPE FROM THE USER'S NOSQL DOCUMENT
 @app.route('/remove_income_type', methods=['POST'])
 def remove_income_type():
     # Get the income type to be removed from the JavaScript request
-    incomeTypeTBR = request.form.get('incomeTypeTBR')
+    incomeTypeTBR = request.json.get('incomeTypeTBR')
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
 
-    dbID = session["user"]
     # Removes the selected income type from the user's NoSQL document
-    col.update_one({"_id" : dbID}, { "$pull" : {"income_types" : incomeTypeTBR}})
+    col.update_one({"_id" : master_user_id}, { "$pull" : {"income_types" : incomeTypeTBR}})
 
     # Resturns a "success" response message to the JavaScript front-end
     return jsonify({'status': 'success'})
 
 ### FUNCTIONALITY TO GET THE RECENT INCOME ENTRIES THE USER HAS ADDED TO THE SQL DATABSE
-@app.route('/get_recent_income', methods = ['GET'])
+@app.route('/get_recent_income', methods = ['POST'])
 def get_recent_income():
     #Retrieves user's id from session
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     recent_income_entries = []
 
     # Queries the database to fetch all income entries for the current user
-    query = f"SELECT income_id, user_id, income_type, amount FROM user_income WHERE user_id = {dbID};"
+    query = f"SELECT income_id, user_id, income_type, amount FROM user_income WHERE user_id = {master_user_id};"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -278,11 +307,13 @@ def get_recent_income():
 @app.route('/delete_income_entry',methods = ['POST'])
 def delete_income_entry():
     # Retreives the unique id of the income entry that is to be deleted
-    incomeEntryTBR = request.form.get('incomeEntryTBR')
+    incomeEntryTBR = request.json.get('incomeEntryTBR')
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     # Crestes and commits a quesry to delete the entry from the database
-    cursor.execute(f"DELETE FROM user_income WHERE income_id = {incomeEntryTBR};")
+    cursor.execute(f"DELETE FROM user_income WHERE income_id = {incomeEntryTBR} AND user_id = {master_user_id};")
     conn.commit()
     conn.close
 
@@ -307,7 +338,8 @@ def add_expense():
     expenseType = request.json.get('expenseType')
     amount = request.json.get('amount')
     date = request.json.get('date')
-    user = session['user']
+    encoded_id = request.json.get('encoded_id')
+    user = decode(encoded_id)
 
     # Create and execute a query that will insert the data into the SQL databse
     conn = mysql.connector.connect(**db_config)
@@ -318,21 +350,24 @@ def add_expense():
     return jsonify({'message' : 'success'})
 
 ###FUNCTIONALITY TO GET ALL THE USER'S EXPENSE TYPES FOR THIER NOSQL DOCUMENT
-@app.route('/get_expense_types', methods = ['GET'])
+@app.route('/get_expense_types', methods = ['POST'])
 def get_expense_types():
-    dbID = session["user"]
-    expenseTypes = col.find_one({"_id": dbID})
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
+    expenseTypes = col.find_one({"_id": master_user_id})
     expenseTypes = expenseTypes['expense_types']
+    print(expenseTypes)
     return jsonify({'types':expenseTypes})
 
 ### FUNCTINALITY TO ADD AN EXPENSE TYPE THE USER'S NOSQL DOCUMENT
 @app.route('/add_expense_type', methods = ['POST'])
 def add_expense_type():
     newExpenseType = request.json.get('newExpenseType')
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
 
     #Gets all expense types in the current user's NoSQL document
-    expenseTypes = col.find_one({"_id": dbID})
+    expenseTypes = col.find_one({"_id": master_user_id})
     expenseTypes = expenseTypes['expense_types'] 
 
     #If expense type exists, return "exists" message to JavaScript front-end
@@ -340,34 +375,35 @@ def add_expense_type():
         return jsonify({'message' : 'exists'})
 
     #If the expense types doesn't exist in the document, it is pushed onto the user's NoSQL document
-    col.update_one({"_id" : dbID}, { "$push" : {"expense_types" : newExpenseType}})
+    col.update_one({"_id" : master_user_id}, { "$push" : {"expense_types" : newExpenseType}})
     return jsonify({'message' : 'success'})
 
 ### FUNCTIONALITY TO REMOVE AN EXPENSE TYPE FROM THE USER'S NOSQL DOCUMENT
 @app.route('/remove_expense_type', methods=['POST'])
 def remove_expense_type():
-    # Get the expense type to be reomved from the JavaScript request
-    
-    expenseTypeTBR = request.form.get('expenseTypeTBR')
+    # Get the expense type to be removed from the JavaScript request    
+    expenseTypeTBR = request.json.get('expenseTypeTBR')
+    encoded_id = request.json.get('encoded_id')
 
     # Get user id from session
-    dbID = session["user"]
-    
+    master_user_id = decode(encoded_id)
+
     # Remove the selected expense type from the user's NOSQL document
-    col.update_one({"_id" : dbID}, { "$pull" : {"expense_types" : expenseTypeTBR}})
+    col.update_one({"_id" : master_user_id}, { "$pull" : {"expense_types" : expenseTypeTBR}})
 
     # Return a "success" response to the JavaScript front-end
     return jsonify({'status': 'success'})
 
 ### FUNCTIONALITY TO GET ALL THE EXPENSE ENTRIES THE USER INPUTTED INTO THE SQL DATABASAE
-@app.route('/get_recent_expenses', methods = ['GET'])
+@app.route('/get_recent_expenses', methods = ['POST'])
 def get_recent_expenses():
     # Gets user's id from session
-    dbID = session["user"]
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     recent_expense_entries = []
 
     # Creates and executes query to return all expense entries inputted by the user
-    query = f"SELECT expense_id, user_id, expense_type, amount FROM user_expenses WHERE user_id = {dbID};"
+    query = f"SELECT expense_id, user_id, expense_type, amount FROM user_expenses WHERE user_id = {master_user_id};"
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -383,12 +419,14 @@ def get_recent_expenses():
 ### FUNCTIONALITY TO DELETE A SPECIFIC EXPENSE ENTRY FROM THE SQL DATASE
 @app.route('/delete_expense_entry',methods = ['POST'])
 def delete_expense_entry():
-    expenseEntryTBR = request.form.get('expenseEntryTBR')
+    expenseEntryTBR = request.json.get('expenseEntryTBR')
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     conn = mysql.connector.connect(**db_config)
     cursor = conn.cursor()
 
     # Creates and executes a query to delete the expense entry selected by the user
-    cursor.execute(f"DELETE FROM user_expenses WHERE expense_id = {expenseEntryTBR};")
+    cursor.execute(f"DELETE FROM user_income WHERE income_id = {expenseEntryTBR} AND user_id = {master_user_id};")
     conn.commit()
     conn.close
 
@@ -406,25 +444,27 @@ def budget():
 
 @app.route('/get_budget_targets', methods = ['GET'])
 def get_budget_targets():
-    dbID = session["user"]
-    budget_targets = col.find_one({"_id": dbID})
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
+    budget_targets = col.find_one({"_id": master_user_id})
     budget_targets = budget_targets['budget']
     return jsonify({'types':budget_targets})
 
 @app.route('/save_budget', methods = ['POST'])
-def save_budget():
-    dbID = session["user"]
+def save_budget():    
     expenseType = request.json.get('expenseType')
     newBudgetAmount = request.json.get('newBudgetAmount')
+    encoded_id = request.json.get('encoded_id')
+    master_user_id = decode(encoded_id)
     newBudgetAmount = int(newBudgetAmount)   
     nested = "budget."+expenseType
 
-    result = col.update_one({'_id': dbID},{'$set': {nested: newBudgetAmount}}, upsert=True)
+    result = col.update_one({'_id': master_user_id},{'$set': {nested: newBudgetAmount}}, upsert=True)
 
     if result.upserted_id:
         return jsonify({'status':'success'})
     else:
-        return jsonify({'status':'success'})
+        return jsonify({'status':'fail'})
 
 ####LOGOUT FUNCTOINALITY
 @app.route('/logout', methods=['POST'])
